Rename k2-pix directory due to hyphen problem in CLI; make local imports; add init
